Version "configurable", with configure and aclocal.m4 files created with autoconf 1.14.1.
For some unknown reason, the file in the previous commits could not be reconfigured (run configure) in Mac OS. It seems that the required configuration files had been recreated using autoconf 1.15, possibly from a version in a linux server. The code in this commit has been created by copying the original maude 2.7.1 code and applying manually the required changes to add the termination checker hook. The configure tool can be successfully be applied in Mac OS for this code.