Maude Alpha77
Dear Maude Abusers,

Bug fixes:
(1)	A bug when reading files that don't end in with a
newline. Maude usually fakes a newline in this case, but if the is an
in/load near the end of the offending file, the flag to fake a newline
will be set before the included file is read and then overwritten by
the handling of the included file before the newline is faked. This is
now fixed by stacking the flag. This bug was found by Peter.

(2)	There was a portability bug in the memory management that
meant Maude would not work unless compiled with g++ 2.95.3. This is
partly fixed by a rewrite of the memory manager. Although it's still
not 100% ANSI compliant it should port to most platforms now.

There are no new features but there are a lot of changes under the
hood:

(1)	It's now compiled with g++ 3.2 and the ANSI compliant
libstdc++ version 3. This may give a performance gain for some
examples. I will probably remove support for the non-ANSI compliant
libstdc++ version 2 in the near future.

(2)	In full AC/ACU matching, element variables are now pushed in
to the bipartite graph problem rather than the Diophantine system
problem in order to reduce the search space. This should help with
Merrill's "killer rules" during search/model checking.

(3)	Iterators used for A/AU/AC/ACU arguments lists; this may
improve performance. SPEED_HACKs removed from A/AU/AC/ACU code.

(4)	More special casing for full AC/ACU matching.

(5)	AC/ACU subpatterns of the form f(X, t) where t is ground and X
is not yet bound now get to use greedy matching even if X occurs
elsewhere in pattern or condition - since subpatterns with this form
can match in at most one way.

(6)	A number of low level changes to the AC/ACU code to avoid
unnecessary copying and unpredicatable branches.

(7)	Matching a variable with extension (i.e. at the top) against
an AC/ACU subject is now a little smarter.

Steven