Maude Alpha84b.1
Dear Maude Abusers,

This version is a slightly patched version of Alpha84b to fix a
critical bug in the AU sort calculation code reported by Jon Haugsand.
A simplified version of his example that exposes the problem is

fmod SORT-INT is
  pr INT .
  sorts Seq Pseq Tree .
  subsorts  Int < Pseq < Seq .
  op ntree : -> Tree [ctor] .
  op tree : Tree Int Tree -> Tree [ctor] .

  op nil : -> Seq .
  op __ : Seq Seq -> Seq    [assoc id: nil ctor] .
  op __ : Pseq Pseq -> Pseq [assoc id: nil ctor] .

  vars M N : Int .
  vars S : Seq .
  vars F G : Tree .

  op downto : Nat Nat -> Seq .
  eq downto(M, N) = if M <= N then  downto(M + 1,N) M else nil fi .

  op itree : Tree Int -> Tree .

  eq itree(ntree,N) = tree(ntree,N,ntree) .
  eq itree(tree(F,M,G),N) = if M >= N then tree(itree(F,N),M,G) else
        tree(F,M,itree(G,N)) fi .

  op seqtotree : Seq -> Tree .
  eq seqtotree(N) = tree(ntree,N,ntree) .
  eq seqtotree(N S) = itree(seqtotree(S),N) .

endfm

red seqtotree(downto(1,6)) .

So far I have had no reports at all about Alpha84b. I would like to
know if it breaks Full Maude in anyway. This new version also reports
itself as 2.1.1 and I would like to release it as such in the next
week or so unless anyone finds a show stopper.

Steven