Maude Alpha84b.2
Dear Maude Abusers,

This release is a slightly patched version of Alpha84b.1 to fix a
critical bug in the operator renaming reported by Paco:

fmod FOO is
  sorts Bar Foo .
  subsorts Bar < Foo .
  op foo : Foo -> Foo .
  op foo : -> Foo .
endfm

fmod BAZ is
  inc FOO * (op foo : Bar -> Bar to bar).
endfm

show all .
show modules .

This new version also reports itself as 2.1.1 and I would like to
release it as such in the next week or so unless anyone finds a show
stopper. I will release a snapshot of my development tree (alpha84c)
with experimental features in the near future.

Steven