Maude Alpha86
Dear Maude Abusers,

This version is a pure alpha release - it is not destined for the
Maude website though of course you may redistribute it under the GPL.
Because of this I have only made binaries for Linux and Darwin.

It is a lot more alpha-like than usual - I'm am releasing it at this
time because of the need to put together a demo for RTA 2005.

Bug Fixes
==========
(1) Loop mode now treats Qid's of the form '